+++ src/main/java/com/vinny/backend/User/domain/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.vinny.backend.Common.domain;
+package com.vinny.backend.User.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
