+++ src/main/java/com/vinny/backend/User/domain/Brand.java
@@ -1,4 +1,4 @@
-package com.vinny.backend.Common.domain;
+package com.vinny.backend.User.domain;
 
 import jakarta.persistence.*;
 import lombok.*;
@@ -8,7 +8,7 @@ import lombok.*;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class Region {
+public class Brand {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -16,4 +16,7 @@ public class Region {
 
     @Column(length = 50, nullable = false)
     private String name;
+
+    @Column(name ="brand_image", nullable = false)
+    private String brandImage;
 }
