+++ src/main/java/com/vinny/backend/User/domain/Region.java
@@ -1,4 +1,4 @@
-package com.vinny.backend.Common.domain;
+package com.vinny.backend.User.domain;
 
 import jakarta.persistence.*;
 import lombok.*;
@@ -8,7 +8,7 @@ import lombok.*;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class Brand {
+public class Region {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -16,7 +16,4 @@ public class Brand {
 
     @Column(length = 50, nullable = false)
     private String name;
-
-    @Column(name ="brand_image", nullable = false)
-    private String brandImage;
 }
