+++ src/main/java/com/vinny/backend/User/domain/SearchLog.java
@@ -1,4 +1,4 @@
-package com.vinny.backend.Common.domain;
+package com.vinny.backend.User.domain;
 
 import jakarta.persistence.*;
 import lombok.*;
