+++ src/main/java/com/vinny/backend/User/domain/mapping/UserVintageItem.java
@@ -1,8 +1,8 @@
-package com.vinny.backend.Common.domain.mapping;
+package com.vinny.backend.User.domain.mapping;
 
-import com.vinny.backend.Common.domain.BaseEntity;
-import com.vinny.backend.Common.domain.Brand;
-import com.vinny.backend.Common.domain.User;
+import com.vinny.backend.User.domain.BaseEntity;
+import com.vinny.backend.User.domain.User;
+import com.vinny.backend.User.domain.VintageItem;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -11,7 +11,7 @@ import lombok.*;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class UserBrand extends BaseEntity {
+public class UserVintageItem extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -22,6 +22,6 @@ public class UserBrand extends BaseEntity {
     private User user;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "brand_id", nullable = false)
-    private Brand brand;
+    @JoinColumn(name = "vintage_item_id", nullable = false)
+    private VintageItem vintageItem;
 }
